+++ Yogscart C#/Assets/Scripts/Power Ups/Projectile.cs
@@ -4,11 +4,10 @@ using UnityEngine;
 
 abstract public class Projectile : MonoBehaviour
 {
-
     protected Vector3 direction;
     protected bool actingShield = false;
 
-    public virtual void Setup(float _direction, bool _actingShield)
+    public virtual void Setup(int _direction, bool _actingShield)
     {
         direction = MathHelper.ZeroYPos(transform.forward * _direction);
         actingShield = _actingShield;
